+++ lib/multitasking.dart
@@ -1,2 +1,3 @@
-export 'src/cancellation.dart';
-export 'src/task.dart';
+export 'src/task/cancellation.dart';
+export 'src/task/errors.dart';
+export 'src/task/task.dart';