+++ lib/synchronization/binary_semaphore.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 import 'counting_semaphore.dart';
+import 'synchronizer.dart';
 
 /// A [BinarySemaphore] is a synchronization primitive with an integer value
 /// restricted to 0 or 1, representing locked (0) or unlocked (1) states.
@@ -13,15 +14,17 @@ import 'counting_semaphore.dart';
 /// If a mutex is locked by a task, it will not block this task. It will
 /// count the number of times it is entered and leaved by task before
 /// releasing.
-class BinarySemaphore {
+class BinarySemaphore implements Synchronizer {
   final CountingSemaphore _lock = CountingSemaphore(0, 1);
 
   /// Acquires a permit from this semaphore.
+  @override
   Future<void> acquire() {
     return _lock.acquire();
   }
 
   /// Releases a permit.
+  @override
   Future<void> release() {
     return _lock.release();
   }
@@ -33,8 +36,9 @@ class BinarySemaphore {
   /// If the timeout expires before the semaphore is unlocked, then returns
   /// `false`.\
   /// If permit was acquired, returns `true`.
+  @override
   @useResult
-  Future<bool> tryAcquire([Duration? timeout]) {
+  Future<bool> tryAcquire(Duration timeout) {
     return _lock.tryAcquire(timeout);
   }
 }