+++ lib/synchronization/synchronizer.dart
@@ -1,0 +1,7 @@
+abstract class Synchronizer {
+  Future<void> acquire();
+
+  Future<void> release();
+
+  Future<bool> tryAcquire(Duration timeout);
+}