+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Added an explanation of the operating principle and internal structure of the task.
 - Changed package description.
+- A unified mechanism for implementing a waiting queue has been added  (`WaitQueue`).
+- Breaking change:The signatures of some synchronization primitives have been changed to consistently use wait queues. This applies to methods that implement waiting with a timeout (`tryWait()`).
 
 ## 2.1.0
 
@@ -15,6 +17,9 @@
 - Breaking change: The functionality of the `CancellationToken` handler has been changed. Performance and usability have been improved, and most importantly, the restriction on linking to a single task has been lifted.
 - Added example: `example_task_cancel_during_sleep.dart`.
 - Added example: `example_binary_semaphore.dart`.
+- Added class: `Synchronizer`.
+- Added class: `ConditionVariable`.
+- Added example: `example_condition_variable.dart`.
 
 ## 2.0.0